LayoutTests/http/tests/security/contentSecurityPolicy/worker-set-timeout-blocked-expected.txt

CONSOLE MESSAGE: Refused to evaluate script because it violates the following Content Security Policy directive: "script-src 'self' 'unsafe-inline'".

ALERT: setTimeout blocked
